requirements.txt
h5py==2.10.0
imageio==2.8.0
joblib==0.15.1
matplotlib==3.1.3
multiprocess==0.70.11.1
numpy==1.18.1
pyparsing==2.4.6
scikit-image==0.16.2
scipy==1.4.1
tensorflow
torchvision
tqdm==4.53.0
